refactor: move Starquestion to UI folder
