+++ src/components/UI/StarQuestion/StarSvg/StarSvg.spec.tsx
