+++ src/components/UI/StarQuestion/StarSvg/index.tsx
